# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FamilyTracker API - Real-time family location tracking
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Profile {
  user_id: string;
  full_name?: string | null;
  avatar_url?: string | null;
  phone?: string | null;
  created_at: string;
  updated_at: string;
}

export interface UpdateProfileRequest {
  full_name?: string;
  avatar_url?: string;
  phone?: string;
}

export interface Family {
  id: string;
  name: string;
  invite_code: string;
  created_by: string;
  created_at: string;
  updated_at: string;
}

export interface CreateFamilyRequest {
  name: string;
}

export interface JoinFamilyRequest {
  invite_code: string;
}

export interface LatestLocation {
  user_id: string;
  latitude: number;
  longitude: number;
  accuracy?: number | null;
  speed?: number | null;
  heading?: number | null;
  is_moving?: boolean | null;
  battery_level?: number | null;
  updated_at: string;
}

export interface UpdateLocationRequest {
  latitude: number;
  longitude: number;
  accuracy?: number | null;
  speed?: number | null;
  heading?: number | null;
  is_moving?: boolean | null;
  battery_level?: number | null;
}

export interface LocationHistoryPoint {
  id: string;
  user_id: string;
  latitude: number;
  longitude: number;
  speed?: number | null;
  is_moving?: boolean | null;
  recorded_at: string;
}

export interface MemberWithLocation {
  user_id: string;
  role: string;
  profile: Profile;
  location?: LatestLocation | null;
}

export interface FamilyWithMembers {
  family: Family;
  members: MemberWithLocation[];
}

export interface Geofence {
  id: string;
  family_id: string;
  name: string;
  latitude: number;
  longitude: number;
  radius: number;
  created_by: string;
  created_at: string;
}

export interface SOSAlert {
  id: string;
  family_id: string;
  user_id: string;
  latitude?: number | null;
  longitude?: number | null;
  message?: string | null;
  is_resolved: boolean;
  created_at: string;
  profile?: Profile | null;
}

export interface FamilyDashboard {
  family: Family;
  members: MemberWithLocation[];
  geofences: Geofence[];
  recent_alerts: SOSAlert[];
}

export interface CreateGeofenceRequest {
  name: string;
  latitude: number;
  longitude: number;
  radius: number;
}

export type MessageMessageType =
  (typeof MessageMessageType)[keyof typeof MessageMessageType];

export const MessageMessageType = {
  text: "text",
  image: "image",
  location_pin: "location_pin",
} as const;

export interface Message {
  id: string;
  family_id: string;
  user_id: string;
  content: string;
  message_type: MessageMessageType;
  image_url?: string | null;
  location_lat?: number | null;
  location_lng?: number | null;
  created_at: string;
  profile?: Profile | null;
}

export type SendMessageRequestMessageType =
  (typeof SendMessageRequestMessageType)[keyof typeof SendMessageRequestMessageType];

export const SendMessageRequestMessageType = {
  text: "text",
  image: "image",
  location_pin: "location_pin",
} as const;

export interface SendMessageRequest {
  content: string;
  message_type?: SendMessageRequestMessageType;
  image_url?: string | null;
  location_lat?: number | null;
  location_lng?: number | null;
}

export interface SendSOSRequest {
  latitude?: number | null;
  longitude?: number | null;
  message?: string | null;
}

export type GetMemberLocationHistoryParams = {
  hours?: number;
};

export type GetFamilyMessagesParams = {
  limit?: number;
  before?: string;
};