# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FamilyTracker API - Real-time family location tracking
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /geofences:
    get:
      operationId: getGeofences
      tags: [geofences]
      summary: Get all geofences for current user's family
      responses:
        "200":
          description: List of geofences
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Geofence"
    post:
      operationId: createGeofence
      tags: [geofences]
      summary: Create a new geofence
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGeofenceRequest"
      responses:
        "201":
          description: Geofence created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Geofence"
  /geofences/{geofenceId}:
    delete:
      operationId: deleteGeofence
      tags: [geofences]
      summary: Delete a geofence
      parameters:
        - name: geofenceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Geofence deleted
  /messages:
    get:
      operationId: getFamilyMessages
      tags: [messages]
      summary: Get family chat messages
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: before
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message to family chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
...
components:
  schemas:
...
    Profile:
      type: object
      properties:
        user_id:
          type: string
        full_name:
          type: string
          nullable: true
        avatar_url:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        created_at:
          type: string
        updated_at:
          type: string
      required:
        - user_id
        - created_at
        - updated_at
...
    Geofence:
      type: object
      properties:
        id:
          type: string
        family_id:
          type: string
        name:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        radius:
          type: number
        created_by:
          type: string
        created_at:
          type: string
      required:
        - id
        - family_id
        - name
        - latitude
        - longitude
        - radius
        - created_by
        - created_at
    CreateGeofenceRequest:
      type: object
      properties:
        name:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        radius:
          type: number
      required:
        - name
        - latitude
        - longitude
        - radius
    Message:
      type: object
      properties:
        id:
          type: string
        family_id:
          type: string
        user_id:
          type: string
        content:
          type: string
        message_type:
          type: string
          enum: [text, image, location_pin]
        image_url:
          type: string
          nullable: true
        location_lat:
          type: number
          nullable: true
        location_lng:
          type: number
          nullable: true
        created_at:
          type: string
        profile:
          $ref: "#/components/schemas/Profile"
          nullable: true
      required:
        - id
        - family_id
        - user_id
        - content
        - message_type
        - created_at
    SendMessageRequest:
      type: object
      properties:
        content:
          type: string
        message_type:
          type: string
          enum: [text, image, location_pin]
          default: text
        image_url:
          type: string
          nullable: true
        location_lat:
          type: number
          nullable: true
        location_lng:
          type: number
          nullable: true
      required:
        - content
...