# lib/api-zod/src/generated/types/createGeofenceRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FamilyTracker API - Real-time family location tracking
 * OpenAPI spec version: 0.1.0
 */

export interface CreateGeofenceRequest {
  name: string;
  latitude: number;
  longitude: number;
  radius: number;
}