# lib/api-zod/src/generated/types/geofence.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FamilyTracker API - Real-time family location tracking
 * OpenAPI spec version: 0.1.0
 */

export interface Geofence {
  id: string;
  family_id: string;
  name: string;
  latitude: number;
  longitude: number;
  radius: number;
  created_by: string;
  created_at: string;
}