# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FamilyTracker API - Real-time family location tracking
 * OpenAPI spec version: 0.1.0
 */
import type { MessageMessageType } from "./messageMessageType";
import type { Profile } from "./profile";

export interface Message {
  id: string;
  family_id: string;
  user_id: string;
  content: string;
  message_type: MessageMessageType;
  image_url?: string | null;
  location_lat?: number | null;
  location_lng?: number | null;
  created_at: string;
  profile?: Profile | null;
}