# lib/api-zod/src/generated/types/sendMessageRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FamilyTracker API - Real-time family location tracking
 * OpenAPI spec version: 0.1.0
 */
import type { SendMessageRequestMessageType } from "./sendMessageRequestMessageType";

export interface SendMessageRequest {
  content: string;
  message_type?: SendMessageRequestMessageType;
  image_url?: string | null;
  location_lat?: number | null;
  location_lng?: number | null;
}